refactor: 배럴 패턴 제거 및 절대경로 추가 적용
