--- src/components/common/Button/index.ts
@@ -1,1 +1,0 @@
-export { default as Button } from './Button';
