--- src/components/common/index.ts
@@ -1,2 +1,0 @@
-export { default as BottomNavBar } from './BottomNavBar';
-export { default as TagList } from './TagList';
