--- src/components/quizComponent/index.ts
@@ -1,7 +1,0 @@
-export { default as QuizAB } from './QuizAB';
-export { default as QuizAns } from './QuizAns';
-export { default as QuizFooter } from './QuizFooter';
-export { default as QuizMul } from './QuizMul';
-export { default as QuizOX } from './QuizOX';
-export { default as QuizWrapper } from './QuizWrapper';
-export { default as QuizRenderer } from './QuizRenderer';
