--- src/pages/Login/index.ts
@@ -1,3 +1,0 @@
-export { default as LoginPage } from './LoginPage';
-export { default as InterestPage } from './InterestPage';
-export { default as CallbackPage } from './CallbackPage';
