--- src/pages/game/index.ts
@@ -1,6 +1,0 @@
-export { default as CreateGame } from './CreateGame';
-export { default as WaitingRoom } from './WaitingRoom';
-export { default as RandomMatch } from './RandomMatch';
-export { default as CodeEntry } from './CodeEntry';
-export { default as Play } from './Play';
-export { default as Result } from './Result';
