--- src/pages/index.ts
@@ -1,4 +1,0 @@
-export { default as GamePage } from './game/GamePage';
-export { default as SearchPage } from './search/SearchPage';
-export { default as MyPage } from './profile/MyPage';
-export * from './main';
