--- src/pages/main/index.ts
@@ -1,2 +1,0 @@
-export { default as MainPage } from './MainPage';
-export { default as QuizCard } from './QuizCard';
