--- src/pages/quiz/index.ts
@@ -1,6 +1,0 @@
-export { default as AbPage } from './AbPage';
-export { default as OxPage } from './OxPage';
-export { default as MultipleChoicePage } from './MultipleChoicePage';
-export { default as EditOxQuizPage } from './EditOxQuizPage';
-export { default as EditMultipleQuizPage } from './EditMultipleQuizPage';
-export { default as EditAbQuizPage } from './EditAbQuizPage';
