--- src/types/index.ts
@@ -1,3 +1,0 @@
-export * from './NavBarType';
-export * from './QuizTypes';
-export * from './CommentTypes';
